moves the ApiTestCase into test.core as there is no real reason for it to be in the loader
